Consolidate close button
