+++ src/app/dim-ui/ClosableContainer.m.scss.d.ts
