--- src/app/dim-ui/CloseButton.m.scss
@@ -1,16 +1,0 @@
-@import '../variables';
-
-.close {
-  width: 16px;
-  height: 16px;
-  display: inline-block;
-  position: absolute;
-  top: 2px;
-  right: calc(2px + var(--item-margin));
-  background-image: url('../../images/close.png');
-  background-size: 16px 16px;
-  background-color: rgba(100, 100, 100, 0.8);
-  &:hover {
-    background-color: $orange;
-  }
-}
