--- src/app/dim-ui/CloseButton.m.scss.d.ts
@@ -1,7 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'close': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
