+++ src/app/loadout-builder/filter/LockedItem.tsx
@@ -1,7 +1,7 @@
+import ClosableContainer from 'app/dim-ui/ClosableContainer';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import DraggableInventoryItem from 'app/inventory/DraggableInventoryItem';
 import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
-import ClosableContainer from 'app/loadout/loadout-ui/ClosableContainer';
 import React from 'react';
 import { LockedItemType } from '../types';
